Added 2018 data from Basel Convention website
Data from the 2018 Basel convention national reports, submitted by 70 out of 189 signatories. 

Note that not all the data is from 2018, some is from 2016, but is included in the 2018 release of the national reports available on the Basel convention website. Extracted from ers.basel.int/ERS-Extended/ExcelFiles/Temp/ERS_Basel_2018-[COUNTRY CODE]-BC-Table4.xlsx . Note the worrying label "Temp" in the url - hopefully this might address that*!

* @GitHub when are you next doing a data backup in Svalbard?