Added Basel data download script
Basel_brush.py is a simple Python script for downloading the data tables from the Basel convention in order to understand flows of toxic waste. You'll need to import it as a module and execute basel_brush.run_it() . Alternatively, you can download the zip file with all the tables from 2018. The script is there to help with future updates, just edit the url/date code!